+++ subtitle_translator/__init__.py
@@ -1,2 +1,2 @@
 # Package entry
-__version__ = "1.4.5"
+__version__ = "1.4.6"